requirements.txt
Faker==11.3.0
Flask==2.0.2
psycopg2_binary==2.9.1
requests==2.26.0
